🎉 Add new tests and remove old integration tests 🎉
- Created `tests/common/mod.rs` to provide common utility functions for test setups.
- Added `tests/feed_validation.rs` for validating the feed with dedicated test cases 👌.
- Introduced `tests/opml_generation.rs` for testing OPML generation functions 🛠️.
- Introduced `tests/opml_parsing.rs` to validate the parsing functionality 🚀.
- Added `tests/report_generation.rs` to validate the formatting of markdown reports 📊.
- Deleted `tests/integration_tests.rs` as it was cluttering the codebase and not adhering to our modular test strategies 🗑️.

The changes add a structured approach to tests by breaking them down into specific modules. This makes it easier to maintain and enhance tests in the future 🤓. The created test files focus on individual functionalities of the `opml_manager`, allowing faster and more precise testing. The removal of the integration tests helps streamline the testing process, increasing overall efficiency. Bye-bye, spaghetti code! 🍝👋
